refactor(tag-field): Simplify tag input handling and integrate contentlet support
