+++ core-web/libs/edit-content/src/lib/fields/dot-edit-content-tag-field/components/tag-field/tag-field.component.html
@@ -1,0 +1,26 @@
+@let variable = $variableName();
+@let hasError = $hasError();
+
+<p-autoComplete
+    #autocomplete
+    [attr.data-testid]="'tag-field-container-' + variable"
+    [id]="'tag-id-' + variable"
+    [inputId]="variable"
+    [suggestions]="$suggestions()"
+    [class.ng-invalid]="hasError"
+    [class.ng-dirty]="hasError"
+    [(ngModel)]="$values"
+    [ngModelOptions]="{ standalone: true }"
+    [multiple]="true"
+    [forceSelection]="false"
+    [unique]="AUTO_COMPLETE_UNIQUE"
+    [minLength]="AUTO_COMPLETE_MIN_LENGTH"
+    [delay]="AUTO_COMPLETE_DELAY"
+    [disabled]="$isDisabled()"
+    (ngModelChange)="onTagsChange($event)"
+    (completeMethod)="onSearch($event)"
+    (keydown.enter)="onEnterKey($event)">
+    <ng-template pTemplate="removetokenicon">
+        <i class="pi pi-times"></i>
+    </ng-template>
+</p-autoComplete>
