+++ core-web/libs/edit-content/src/lib/fields/dot-edit-content-tag-field/dot-edit-content-tag-field.component.html
@@ -1,32 +1,17 @@
 @let field = $field();
-@let variable = field.variable;
+@let showLabel = $showLabel();
 
 <dot-card-field [hasError]="hasError">
-    <ng-content />
+    @if (showLabel) {
+        <dot-card-field-label [variableName]="field.variable" [isRequired]="isRequired">
+            {{ field.name }}
+        </dot-card-field-label>
+    }
     <dot-card-field-content>
-        <p-autoComplete
-            #autocomplete
-            [attr.data-testid]="'tag-field-container-' + variable"
-            [id]="'tag-id-' + variable"
-            [inputId]="variable"
-            [suggestions]="$suggestions()"
-            [class.ng-invalid]="hasError"
-            [class.ng-dirty]="hasError"
-            [(ngModel)]="$values"
-            [ngModelOptions]="{ standalone: true }"
-            [multiple]="true"
-            [forceSelection]="false"
-            [unique]="AUTO_COMPLETE_UNIQUE"
-            [minLength]="AUTO_COMPLETE_MIN_LENGTH"
-            [delay]="AUTO_COMPLETE_DELAY"
-            [disabled]="$disabled()"
-            (ngModelChange)="onTagsChange($event)"
-            (completeMethod)="onSearch($event)"
-            (keydown.enter)="onEnterKey($event)">
-            <ng-template pTemplate="removetokenicon">
-                <i class="pi pi-times"></i>
-            </ng-template>
-        </p-autoComplete>
+        <dot-tag-field
+            [formControlName]="field.variable"
+            [variableName]="field.variable"
+            [hasError]="hasError" />
     </dot-card-field-content>
     <dot-card-field-footer>
         @if (hasError) {
